admin_panel/requirements.txt
django==4.0.4
django-split-settings==1.1.0
flake8==4.0
flake8-isort==4.2.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
uwsgi==2.0.20
